LoraDataHandler: rename all header files to *.h
